基本完成权限管理菜单，待完善代码方面 权限管理-菜单v-0.0.2

+++ Front/src/views/Permission/Menu/Menu.vue
@@ -1,8 +1,33 @@
 <template>
   <div class="permission-menu-wrapper">
     <div class="conditions-wrapper">
-      <el-button size='small' type='success' @click='addItem'>新增用户</el-button>
+      <el-button size='small' type='success' @click='addItem' class="mgr-20">新增用户</el-button>
+      <el-date-picker
+        class="mgr-20"
+        v-model="timeRange"
+        type="daterange"
+        range-separator="至"
+        start-placeholder="开始日期"
+        end-placeholder="结束日期"
+        size="small"
+        @change="timePickerChange"
+        :picker-options='pickerOptions'
+      />
+      <el-select v-if='$store.state.userInfo.role === "SUPERADMIN"' size="small" clearable v-model="params.roleLevel" @change='roleLevelChange' placeholder="请选择用户权限级别">
+        <el-option
+          v-for="item in roleEnum"
+          :key="item.value"
+          :label="item.label"
+          :value="item.value"
+        />
+      </el-select>
       <span class="auto-flex"></span>
+      <SearchBar
+        class="search-bar"
+        placeholder="Name/Mobile"
+        v-model="params.keyword"
+        @sureKeyword="sureKeyword"
+      />
     </div>
     <div class="table-wrapper">
       <el-table
@@ -24,10 +49,10 @@
           </template>
         </el-table-column>
         <el-table-column prop="roleDesc" label="Role" align='center' />
-        <el-table-column label="Opertion" align='center'>
+        <el-table-column label="Opertion" align='center' v-if='$hasPermission("PERMISSIONMENUEDIT,PERMISSIONMENUDELETE")'>
           <template slot-scope="scope">
-            <el-button size='mini' type='primary' @click.native='handleEdit(scope.row._id)'>edit</el-button>
-            <el-button size='mini' type='danger' @click.native='handleDelete(scope.row._id)'>delete</el-button>
+            <el-button size='mini' type='primary' @click.native='handleEdit(scope.row._id)' v-permission="'PERMISSIONMENUEDIT'">edit</el-button>
+            <el-button size='mini' type='danger' :disabled='scope.row.role === "SUPERADMIN"' @click.native='handleDelete(scope.row._id)' v-permission="'PERMISSIONMENUDELETE'">delete</el-button>
           </template>
         </el-table-column>
       </el-table>
@@ -43,11 +68,13 @@
         :total="count">
       </el-pagination>
     </div>
+    <!-- 编辑弹窗 -->
     <EditPermissionDialog
       :editPermissionDialog='editPermissionDialog'
       :id='id'
       @closePermissionDialog='closePermissionDialog'
     />
+    <!-- 新增弹窗 -->
     <AddPermissionDialog
       :addPermissionDialog='addPermissionDialog'
       @closeAddPermissionDialog='closeAddPermissionDialog'
@@ -71,19 +98,54 @@ export default {
     return {
       userInfos: null,
       count: 0,
+      theTimeRange: null,
       params: {
         page: 1,
-        size: 20
+        size: 20,
+        startTime: null,
+        endTime: null,
+        keyword: null,
+        roleLevel: null
+      },
+      pickerOptions: {
+        disabledDate (date) {
+          return date > Date.now();
+        }
       },
       id: null,
       editPermissionDialog: false,
       addPermissionDialog: false
     };
   },
-  computed: {},
+  computed: {
+    timeRange: {
+      get () {
+        return this.theTimeRange;
+      },
+      set (val) {
+        if (!val) {
+          this.params.startTime = null;
+          this.params.ednTime = null;
+          this.theTimeRange = null;
+        } else {
+          this.theTimeRange = val;
+          this.params.startTime = $formDate(val[0], 'yyyy-MM-dd');
+          this.params.endTime = $formDate(val[1], 'yyyy-MM-dd');
+        }
+      }
+    },
+    roleEnum () {
+      return [
+        { value: 'COMMON', label: '普通用户' },
+        { value: 'ADMIN', label: '管理员' },
+        { value: 'SUPERADMIN', label: '超级管理员' }
+      ];
+    }
+  },
   methods: {
-    getAllUserInfo (params) {
-      this.$axios.get('/getAllUserInfo', { params })
+    // 获取用户列表
+    getAllUserInfo () {
+      this.$axios.get('/getAllUserInfo', { params: this.params })
         .then(res => {
           const { result, status } = res.data;
           if (status === 0) {
@@ -93,41 +155,86 @@ export default {
           }
         });
     },
+    // 删除单个用户
     async deleteSingleUserById (params) {
       return await this.$axios.get('/deleteSingleUserById', { params });
     },
+    // 分页器 size 改变
     handleSizeChange (val) {
       this.params.size = val;
-      this.getAllUserInfo(this.params);
+      this.getAllUserInfo();
     },
+    // 分页器 page 改变
     handleCurrentChange (val) {
       this.params.page = val;
-      this.getAllUserInfo(this.params);
+      this.getAllUserInfo();
     },
+    // 编辑用户事件
     handleEdit (id) {
       this.id = id;
       this.editPermissionDialog = true;
     },
+    // 删除用户事件
     handleDelete (_id) {
-      this.deleteSingleUserById({_id})
+      this.$confirm('此操作将永久删除该条数据, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        this.deleteSingleUserById({_id})
         .then((res) => {
           const { status } = res.data;
           if (status === 0) {
             this.getAllUserInfo();
           }
         });
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        });          
+      });
     },
+    // 编辑弹窗回调
     closePermissionDialog () {
       this.editPermissionDialog = false;
     },
+    // 新增用户
     addItem () {
       this.addPermissionDialog = true;
     },
+    // 新增弹窗取消回调
     closeAddPermissionDialog () {
       this.addPermissionDialog = false;
     },
+    // 新增弹窗确定回调
     surePermissionDialog () {
       this.addPermissionDialog = false;
+      this.getAllUserInfo();
+    },
+    // datePicker 改变事件
+    timePickerChange () {
+      this.params.keyword = null;
+      this.params.roleLevel = null;
+      this.getAllUserInfo();
+    },
+    // 级别选择框改变
+    roleLevelChange (val) {
+      this.params.startTime = null;
+      this.params.endTime = null;
+      this.theTimeRange = null;
+      this.params.keyword = null;
+      if (!val) {
+        this.params.roleLevel = null;
+      }
+      this.getAllUserInfo();
+    },
+    // 关键字搜索改变
+    sureKeyword () {
+      this.params.startTime = null;
+      this.params.endTime = null;
+      this.theTimeRange = null;
+      this.params.roleLevel = null;
       this.getAllUserInfo();
     }
   },
@@ -137,7 +244,7 @@ export default {
     }
   },
   created () {
-    this.getAllUserInfo(this.params);
+    this.getAllUserInfo();
   },
   mounted () {},
   watch: {}
